Ultimate Moving Average Xing v.3.6 by Ferretz
INTRODUCTION
Free, (newly) open-source and donation-based customizable (Types and Ratios) MA Crossing bot with the following extra features:

Many Moving Averages to chose from for defining your own rules
Different threshold if the move is profitable (based on the last operation)
Buy at market price or at defined limit price (% of market price)
MA Delta and thresholds are rendered on graph for easy finetuning of the MA parameters (Type, Short, Long, Profitable threshold, Unprofitable threshold) to get the desired behaviour for a specific market
It will perform only 1 BUY or SELL operation at the time, wait until conditions are met to perform the opposite operation and wait again until the next good moment (or if the situation is getting too bad).

Please use this topic for questions and to share your settings and results: 🐭 Ultimate Moving Average Xing
HOW IT WORKS
There are 3 modes to chose from which will totally change the decisions of the bot and when it will trigger the BUY / SELL orders. It's important to understand the differences between to set it up correctly:

Threshold High/Low Mode (aka Basic Mode)

By default it will use only the SHORT and MID Moving Averages. Based on the difference of these 2 it calculates the delta and when higher than the upper threshold it will place a BUY order. When it goes below the lower threshold, it will place a SELL order.

Setting 0 as threshold values will trigger a corresponding order as soon as the Short MA and the Medium MA are crossing over and sell as soon as crossing back below.

Kick & Hold Mode

It's possible to add a 3rd Moving Average, the LONG one. When a LONG Moving Average is used, it's recommended to switch to the Kick & Hold Mode instead of using the High/Low Threshold Mode as the 3rd MA will have no effect otherwise (apart being rendered on the graph).

Rules of Kick & Hold:

To place a BUY order, the SHORT/MID delta must be over the High Threshold (same as standard mode)
As long as MID/LONG delta is over the upper threshold, then it will Hold (so it won't SELL) even if SHORT/MID crosses below 0
It will SELL as soon as SHORT/MID and MID/LONG deltas are passing below the upper threshold
So in this mode, the lower threshold is not used as all and to better help you in fine-tuning the 3 MA's, an indicator about the Kick and Hold status is indicated at the top of the chart.

I have changed that in v3.0 because it seems to make much more sense to me and it's a very similar approach to what I use in 🌙 Moon Catcher.

Please keep in mind that it's not an out-of-the-box / ready to use strategy with which you should just click the "Go live" button but rather a nice tool to test different approaches on a given market. It may achieve nice results once configured properly but do not let it drive real currencies unless you're sure of what you are doing :)

MACD

MACD stands for convergence / divergence. The best is probably that you start by some lecture: http://www.investopedia.com/walkthrough/forex/intermediate/level4/macd.aspx

In this mode it will use SHORT and MID and calculate the divergence between these 2. Then it will calculate the MACD signal, commonly a value of SMA 9 is used for the signal which is the default value (if you let the 3rd MA to NONE) however you're able to override this setting and setting manually any kind of MA or period length for the signal line by defining the 3rd MA.

SUPPORTED MOVING AVERAGES
SMA - Simple Moving Average

EMA - Exponential Moving Average

WMA - Weighted Moving Average

DEMA - Double Exponential Moving Average

TEMA - Triple Exponential Moving Average

TRIMA - Triangular Moving Average

KAMA - Kaufman Adaptive Moving Average

MAMA/FAMA - MESA Adaptive Moving Average

HMA - Hull Moving Average

OUTPUT CHART
EEMAX

More or less everything is rendered on the graph so that it's easier to understand the decisions it takes and to adjust it.

On top there is the Kick/Hold indicator. When kicking it places a BUY order, when not holding anymore, it places a SELL order.

As explained under the HOW IT WORKS section, depending if you are in Basic Mode with 2 Moving Averages or in Kick and Hold Mode with 3 Moving Averages, the rules slightly differ.

You can optionally display the differential graphs. A gray line shows the middle and the current thresholds and in blue the current state. When running 3 Moving Averages you have 2 delta graphs rendered.

On the bottom the profitability graph which comes from 🌙 Moon Catcher. During a BUY operation, it indicates if it is a profitable move compared to the price it was bought. In red if the current price is below the purchased price, in green if it's over the purchased price and in dark green in case it reaches 25% of profit. Only the very end of each bar counts since it indicates if it was able to SELL at a higher price or not.

OUTPUT LOGS
EEMAX

Like several other parts of this bot, it directly comes from 🌙 Moon Catcher. Should contain only relevant and hopefully self-explaining indications :)

INPUT PARAMETERS
MA_TYPE_SHORT/ MA_SHORT / MA_TYPE_LONG / MA_LONG
Allows to chose the desired MA Type ('SMA','EMA','WMA','DEMA','TEMA','TRIVA','KAMA','MAMA') and Ratio (5/8, 10/21, ...)

CROSSING_MODE
Allows choosing what triggers the buy/sell orders.

The default is crossing above the cross + threshold or below the cross - threshold.

If you use 3 moving averages, it's possible to use the kick and hold mode and a MACD approach is also available (will use SMA 9 periods signal if the 3rd MA is set to none, otherwise lets you select the MA Type and period manually).

PROFITABLE_THRESHOLD / UNPROFITABLE_THRESHOLD
Threshold at which to BUY or SELL whenever the operation is profitable (SELLING at higher price than it was bought or BUYING at lower price than it was sold) or not.

BUY_ORDER_SIZE / SELL_ORDER_SIZE
Percentage of Currenty or Asset to place in each order (there is only 1 order at the time)

ORDER_TYPE
Allows chosing between: Market, Limit and Iceberg order types

BUY_PRICE
Used for Limit mode to define at which value to place the orders. If 1.002 is set as value then it will place an order to buy at slightly higher price than the market price. On the other end, setting 0.998 will try to buy at lower price than the market's price (and potentially not be filled)

STOP_LOSS
If the last operation's profit goes below this value then it will be cancelled by placing an opposite operation

PLOT_DIFF
Display the MA Delta Chart

EXTRA
This is a donation-based bot so if you enjoy it, have fun, make profit or have feature requests, your donations are appreciated:

BTCW

18CKm9wRXzQAu7VbS9wSE1dZJby1Tirp6m