Percentile Nearest Rank
It just sorts all the values of an array and finds the value closest to your percent input (decimal). 0.5 or 50% is the median. This draws a line down the visual middle. PNR could be used to find relative highs or lows, or to find thresholds in indicator data (0.9,0.1,0.333,0.687etc). If you find any errors please post a fix.